feat: added versioning to the api
versioned all the api routes by seperating the version 1 routes in
/server/src/routes/versions/v1.api.js which exports a router containing
v1 routes. i.e. /v1/launches/.. type of endpoints. this makes our
api future proof as if we had any breaking changes in the future for
our api we could launch a version 2 and still the aplications using
our api won't break as they will be running on version 1 so they have
time to migrate to version 2 if need be.

signed-off-by: Nisarg Patel(patelnjisarg118@gmail.com).